+++ src/main/java/com/mobble/mobbleserver/application/club/core/service/ClubQueryService.java
@@ -2,6 +2,7 @@ package com.mobble.mobbleserver.application.club.core.service;
 
 import com.mobble.mobbleserver.application.club.core.port.provided.ClubQueryPort;
 import com.mobble.mobbleserver.application.club.core.port.required.ClubReadPort;
+import com.mobble.mobbleserver.application.clubMember.port.required.ClubMemberReadPort;
 import com.mobble.mobbleserver.application.member.port.required.MemberReadPort;
 import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.domain.member.Member;
@@ -17,9 +18,8 @@ import com.mobble.mobbleserver.refactor.club.ageGroup.repository.AgeGroupReposit
 import com.mobble.mobbleserver.refactor.club.clubGround.repository.ClubGroundRepository;
 import com.mobble.mobbleserver.infrastructure.web.club.core.dto.request.ClubSearchRequestDto;
 import com.mobble.mobbleserver.infrastructure.web.club.core.dto.response.ClubSummaryDto;
-import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
-import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMemberRole;
-import com.mobble.mobbleserver.refactor.clubMember.repository.ClubMemberRepository;
+import com.mobble.mobbleserver.domain.clubMember.ClubMember;
+import com.mobble.mobbleserver.domain.clubMember.ClubMemberRole;
 import com.mobble.mobbleserver.refactor.ground.dto.response.GroundResponseDto;
 import com.mobble.mobbleserver.refactor.ground.entity.Ground;
 import com.mobble.mobbleserver.refactor.ground.repository.GroundRepository;
@@ -37,8 +37,8 @@ public class ClubQueryService implements ClubQueryPort {
 
     private final ClubReadPort clubReadPort;
     private final MemberReadPort memberReadPort;
+    private final ClubMemberReadPort clubMemberReadPort;
 
-    private final ClubMemberRepository clubMemberRepository;
     private final AgeGroupRepository ageGroupRepository;
     private final GroundRepository groundRepository;
     private final ClubGroundRepository clubGroundRepository;
@@ -46,8 +46,9 @@ public class ClubQueryService implements ClubQueryPort {
     @Override
     public ClubResponseDto findClubById(Long clubId, Long memberId) {
         Club club = findClubByClubIdOrThrow(clubId);
-        ClubMember leader = clubMemberRepository
+        ClubMember leader = clubMemberReadPort
                 .findByClubIdAndClubMemberRole(clubId, ClubMemberRole.LEADER).get();
+
         String leaderName = leader.getMember().getName();
         Member member = findMemberByMemberIdOrThrow(memberId);
