+++ src/main/java/com/mobble/mobbleserver/application/comment/service/CommentModifyService.java
@@ -1,21 +1,22 @@
 package com.mobble.mobbleserver.application.comment.service;
 
+import com.mobble.mobbleserver.application.clubMember.port.required.ClubMemberReadPort;
 import com.mobble.mobbleserver.application.comment.port.provided.CommentCreatePort;
 import com.mobble.mobbleserver.application.comment.port.provided.CommentDeletePort;
 import com.mobble.mobbleserver.application.comment.port.provided.CommentUpdatePort;
 import com.mobble.mobbleserver.application.comment.port.required.CommentReadPort;
 import com.mobble.mobbleserver.application.comment.port.required.CommentWritePort;
+import com.mobble.mobbleserver.domain.club.core.Club;
+import com.mobble.mobbleserver.domain.clubMember.ClubMember;
 import com.mobble.mobbleserver.domain.comment.Comment;
+import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
+import com.mobble.mobbleserver.global.exception.errorCode.club.ClubMemberErrorCode;
 import com.mobble.mobbleserver.global.exception.errorCode.comment.CommentErrorCode;
 import com.mobble.mobbleserver.infrastructure.web.comment.dto.request.CommentRequestDto;
 import com.mobble.mobbleserver.refactor.article.entity.Article;
 import com.mobble.mobbleserver.refactor.article.validator.ArticleValidator;
-import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.refactor.club.policy.ClubPermissionPolicy;
-import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
-import com.mobble.mobbleserver.refactor.clubMember.validator.ClubMemberValidator;
-import com.mobble.mobbleserver.domain.member.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -25,17 +26,17 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class CommentModifyService implements CommentCreatePort, CommentUpdatePort, CommentDeletePort {
 
-    private final ClubMemberValidator clubMemberValidator;
     private final ArticleValidator articleValidator;
 
     private final CommentWritePort commentWritePort;
     private final CommentReadPort commentReadPort;
+    private final ClubMemberReadPort clubMemberReadPort;
 
     @Override
     public Comment createRootComment(Long memberId, Long articleId, CommentRequestDto dto) {
         Article article = articleValidator.findArticleByArticleIdOrThrow(articleId);
         Long clubId = article.getClub().getId();
-        ClubMember clubMember = clubMemberValidator.findClubMemberByClubIdAndMemberIdOrThrow(clubId, memberId);
+        ClubMember clubMember = findClubMemberByClubIdAndMemberIdOrThrow(clubId, memberId);
 
         Member member = clubMember.getMember();
         Comment comment = dto.toEntity(member, article);
@@ -52,7 +53,7 @@ public class CommentModifyService implements CommentCreatePort, CommentUpdatePor
     ) {
         Article article = articleValidator.findArticleByArticleIdOrThrow(articleId);
         Long clubId = article.getClub().getId();
-        ClubMember clubMember = clubMemberValidator.findClubMemberByClubIdAndMemberIdOrThrow(clubId, memberId);
+        ClubMember clubMember = findClubMemberByClubIdAndMemberIdOrThrow(clubId, memberId);
         Comment parentComment = commentReadPort.findById(parentCommentId).orElseThrow();
 
         Member member = clubMember.getMember();
@@ -70,7 +71,7 @@ public class CommentModifyService implements CommentCreatePort, CommentUpdatePor
     ) {
         Article article = articleValidator.findArticleByArticleIdOrThrow(articleId);
         Club club = article.getClub();
-        ClubMember clubMember = clubMemberValidator.findClubMemberByClubIdAndMemberIdOrThrow(club.getId(), memberId);
+        ClubMember clubMember = findClubMemberByClubIdAndMemberIdOrThrow(club.getId(), memberId);
         Member member = clubMember.getMember();
         Comment comment = commentReadPort.findByIdAndMemberId(commentId, member.getId()).orElseThrow();
 
@@ -83,7 +84,7 @@ public class CommentModifyService implements CommentCreatePort, CommentUpdatePor
     public void deleteComment(Long articleId, Long commentId, Long memberId) {
         Article article = articleValidator.findArticleByArticleIdOrThrow(articleId);
         Club club = article.getClub();
-        ClubMember clubMember = clubMemberValidator.findClubMemberByClubIdAndMemberIdOrThrow(club.getId(), memberId);
+        ClubMember clubMember = findClubMemberByClubIdAndMemberIdOrThrow(club.getId(), memberId);
 
         Comment comment;
 
@@ -100,5 +101,10 @@ public class CommentModifyService implements CommentCreatePort, CommentUpdatePor
 
     private void validateCommentByArticleIdOrThrow(Comment comment, Long articleId) {
         if (!comment.getArticle().getId().equals(articleId)) throw new DomainException(CommentErrorCode.ARTICLE_REQUIRED);
+    }
+
+    private ClubMember findClubMemberByClubIdAndMemberIdOrThrow(Long clubId, Long memberId) {
+        return clubMemberReadPort.findClubMemberByClubIdAndMemberId(clubId, memberId)
+                .orElseThrow(() -> new DomainException(ClubMemberErrorCode.NOT_JOINED_CLUB));
     }
 }
