+++ src/main/java/com/mobble/mobbleserver/application/meeting/service/MeetingModifyService.java
@@ -1,18 +1,19 @@
 package com.mobble.mobbleserver.application.meeting.service;
 
+import com.mobble.mobbleserver.application.clubMember.port.required.ClubMemberReadPort;
 import com.mobble.mobbleserver.application.meeting.port.provided.MeetingCreatePort;
 import com.mobble.mobbleserver.application.meeting.port.provided.MeetingDeletePort;
 import com.mobble.mobbleserver.application.meeting.port.provided.MeetingUpdatePort;
 import com.mobble.mobbleserver.application.meeting.port.required.MeetingReadPort;
 import com.mobble.mobbleserver.application.meeting.port.required.MeetingWritePort;
+import com.mobble.mobbleserver.domain.clubMember.ClubMember;
 import com.mobble.mobbleserver.domain.meeting.Meeting;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
+import com.mobble.mobbleserver.global.exception.errorCode.club.ClubMemberErrorCode;
 import com.mobble.mobbleserver.global.exception.errorCode.meeting.MeetingErrorCode;
 import com.mobble.mobbleserver.infrastructure.web.meeting.dto.request.MeetingRequestDto;
 import com.mobble.mobbleserver.infrastructure.web.meeting.dto.request.MeetingUpdateRequestDto;
 import com.mobble.mobbleserver.refactor.club.policy.ClubPermissionPolicy;
-import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
-import com.mobble.mobbleserver.refactor.clubMember.validator.ClubMemberValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -22,14 +23,13 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional
 public class MeetingModifyService implements MeetingCreatePort, MeetingUpdatePort, MeetingDeletePort {
 
-    private final ClubMemberValidator clubMemberValidator;
-
     private final MeetingWritePort meetingWritePort;
     private final MeetingReadPort meetingReadPort;
+    private final ClubMemberReadPort clubMemberReadPort;
 
     @Override
     public Meeting createMeeting(Long memberId, Long clubId, MeetingRequestDto dto) {
-        ClubMember hostMember = clubMemberValidator.findClubMemberByClubIdAndMemberIdOrThrow(clubId, memberId);
+        ClubMember hostMember = findClubMemberByClubIdAndMemberIdOrThrow(clubId, memberId);
         ClubPermissionPolicy.validateLeaderOrManagerOrThrow(hostMember);
 
         Meeting meeting = dto.toEntity(hostMember);
@@ -39,7 +39,7 @@ public class MeetingModifyService implements MeetingCreatePort, MeetingUpdatePor
 
     @Override
     public Meeting updateMeeting(Long memberId, Long clubId, Long meetingId, MeetingUpdateRequestDto dto) {
-        ClubMember clubMember = clubMemberValidator.findClubMemberByClubIdAndMemberIdOrThrow(clubId, memberId);
+        ClubMember clubMember = findClubMemberByClubIdAndMemberIdOrThrow(clubId, memberId);
         ClubPermissionPolicy.validateLeaderOrManagerOrThrow(clubMember);
         Meeting meeting = findMeetingByMeetingIdOrThrow(meetingId);
 
@@ -55,7 +55,7 @@ public class MeetingModifyService implements MeetingCreatePort, MeetingUpdatePor
 
     @Override
     public void deleteMeeting(Long memberId, Long clubId, Long meetingId) {
-        ClubMember clubMember = clubMemberValidator.findClubMemberByClubIdAndMemberIdOrThrow(clubId, memberId);
+        ClubMember clubMember = findClubMemberByClubIdAndMemberIdOrThrow(clubId, memberId);
         ClubPermissionPolicy.validateLeaderOrManagerOrThrow(clubMember);
         Meeting meeting = findMeetingByMeetingIdOrThrow(meetingId);
 
@@ -65,5 +65,10 @@ public class MeetingModifyService implements MeetingCreatePort, MeetingUpdatePor
     private Meeting findMeetingByMeetingIdOrThrow(Long meetingId) {
         return meetingReadPort.findById(meetingId)
                 .orElseThrow(() -> new DomainException(MeetingErrorCode.NOT_FOUND_MEETING));
+    }
+
+    private ClubMember findClubMemberByClubIdAndMemberIdOrThrow(Long clubId, Long memberId) {
+        return clubMemberReadPort.findClubMemberByClubIdAndMemberId(clubId, memberId)
+                .orElseThrow(() -> new DomainException(ClubMemberErrorCode.NOT_JOINED_CLUB));
     }
 }
