+++ src/main/java/com/mobble/mobbleserver/application/meeting/service/MeetingQueryService.java
@@ -1,9 +1,9 @@
 package com.mobble.mobbleserver.application.meeting.service;
 
+import com.mobble.mobbleserver.application.clubMember.port.required.ClubMemberReadPort;
 import com.mobble.mobbleserver.application.meeting.port.provided.MeetingQueryPort;
 import com.mobble.mobbleserver.application.meeting.port.required.MeetingReadPort;
 import com.mobble.mobbleserver.domain.meeting.Meeting;
-import com.mobble.mobbleserver.refactor.clubMember.validator.ClubMemberValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,13 +15,12 @@ import java.util.List;
 @Transactional(readOnly = true)
 public class MeetingQueryService implements MeetingQueryPort {
 
-    private final ClubMemberValidator clubMemberValidator;
-
     private final MeetingReadPort meetingReadPort;
+    private final ClubMemberReadPort clubMemberReadPort;
 
     @Override
     public List<Meeting> findMeetingsByClubId(Long memberId, Long clubId) {
-        clubMemberValidator.findClubMemberByClubIdAndMemberIdOrThrow(clubId, memberId);
+        clubMemberReadPort.findClubMemberByClubIdAndMemberId(clubId, memberId);
 
         return meetingReadPort.findByClubMember_Club_Id(clubId);
     }
