+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/meeting/dto/request/MeetingRequestDto.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.infrastructure.web.meeting.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
+import com.mobble.mobbleserver.domain.clubMember.ClubMember;
 import com.mobble.mobbleserver.domain.meeting.Meeting;
 import com.mobble.mobbleserver.domain.meeting.MeetingType;
