+++ src/main/java/com/mobble/mobbleserver/refactor/club/policy/ClubPermissionPolicy.java
@@ -1,9 +1,9 @@
 package com.mobble.mobbleserver.refactor.club.policy;
 
+import com.mobble.mobbleserver.domain.clubMember.ClubMember;
+import com.mobble.mobbleserver.domain.clubMember.ClubMemberRole;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.security.SecurityErrorCode;
-import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
-import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMemberRole;
 
 public final class ClubPermissionPolicy {
