+++ src/test/java/com/mobble/mobbleserver/support/fixture/meeting/MeetingTestFixture.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.support.fixture.meeting;
 
-import com.mobble.mobbleserver.refactor.clubMember.entity.ClubMember;
+import com.mobble.mobbleserver.domain.clubMember.ClubMember;
 import com.mobble.mobbleserver.domain.meeting.Meeting;
 import com.mobble.mobbleserver.domain.meeting.MeetingType;
